API review: move avocado.linux.distro to avocado.utils.distro
This is a utility module with no dependency on any other Avocado
module, so a good candidate for the utils namespace.
